// src/app/models/veiculo.ts
import { Optional } from "@angular/core";

export interface Veiculo{
    id?: any;
    cor: String;
    placa: String;
    fabricante: String;
    combustivel: String;
    modelo: String;
    nacionalidade: String; 
    categoria:String
    ano: String;
    diaria:String;
    chassi:String;
   
   


}